Changed int to size_t

+++ tests/accessor/host_accessor_api_common.h
@@ -111,7 +111,7 @@ class run_api_tests {
       auto offset_id =
           util::get_cts_object::id<dims>::get(offset, offset, offset);
       std::remove_const_t<T> data[buff_size];
-      for (int i = 0; i < buff_size; i++)
+      for (size_t i = 0; i < buff_size; i++)
         data[i] = value_operations::init<T>(i);
       bool res = false;
       {
